refactor: Improve HF adapter API
