+++ adapters/huggingface/errors.go
@@ -14,6 +14,6 @@ var (
 )
 
 // Wrap wraps an error with additional context
-func Wrap(err error, wrapErr error, msg string) error {
+func wrap(err error, wrapErr error, msg string) error {
 	return errors.New(msg + ": " + wrapErr.Error() + ": " + err.Error())
 }
